Move file collection fingerprinting infrastructure to :execution
