+++ subprojects/core/src/main/java/org/gradle/internal/fingerprint/AbsolutePathInputNormalizer.java
@@ -17,6 +17,7 @@
 package org.gradle.internal.fingerprint;
 
 import org.gradle.api.tasks.FileNormalizer;
+import org.gradle.internal.execution.fingerprint.DirectorySensitiveNormalizer;
 
 /**
  * {@link FileNormalizer} that uses absolute paths for input files. The default.
