+++ subprojects/core/src/main/java/org/gradle/internal/fingerprint/RelativePathInputNormalizer.java
@@ -17,6 +17,7 @@
 package org.gradle.internal.fingerprint;
 
 import org.gradle.api.tasks.FileNormalizer;
+import org.gradle.internal.execution.fingerprint.DirectorySensitiveNormalizer;
 
 /**
  * {@link FileNormalizer} that uses the location of files in a hierarchy as normalized paths.
