+++ subprojects/core/src/main/java/org/gradle/internal/fingerprint/classpath/ClasspathFingerprinter.java
@@ -16,8 +16,8 @@
 
 package org.gradle.internal.fingerprint.classpath;
 
+import org.gradle.internal.execution.fingerprint.FileCollectionFingerprinter;
 import org.gradle.internal.fingerprint.FileCollectionFingerprint;
-import org.gradle.internal.fingerprint.FileCollectionFingerprinter;
 import org.gradle.internal.fingerprint.impl.RelativePathFileCollectionFingerprinter;
 
 /**
