+++ subprojects/core/src/main/java/org/gradle/internal/fingerprint/impl/IgnoredPathFileCollectionFingerprinter.java
@@ -17,7 +17,7 @@
 package org.gradle.internal.fingerprint.impl;
 
 import org.gradle.api.tasks.FileNormalizer;
-import org.gradle.internal.fingerprint.FileCollectionSnapshotter;
+import org.gradle.internal.execution.fingerprint.FileCollectionSnapshotter;
 import org.gradle.internal.fingerprint.IgnoredPathInputNormalizer;
 
 public class IgnoredPathFileCollectionFingerprinter extends AbstractFileCollectionFingerprinter {
