+++ subprojects/core/src/main/java/org/gradle/internal/fingerprint/impl/RelativePathFileCollectionFingerprinter.java
@@ -18,8 +18,8 @@ package org.gradle.internal.fingerprint.impl;
 
 import org.gradle.api.internal.cache.StringInterner;
 import org.gradle.api.tasks.FileNormalizer;
+import org.gradle.internal.execution.fingerprint.FileCollectionSnapshotter;
 import org.gradle.internal.fingerprint.DirectorySensitivity;
-import org.gradle.internal.fingerprint.FileCollectionSnapshotter;
 import org.gradle.internal.fingerprint.RelativePathInputNormalizer;
 
 public class RelativePathFileCollectionFingerprinter extends AbstractFileCollectionFingerprinter {
