+++ subprojects/execution/src/main/java/org/gradle/internal/execution/fingerprint/FileCollectionFingerprinterRegistry.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.fingerprint;
+package org.gradle.internal.execution.fingerprint;
 
 public interface FileCollectionFingerprinterRegistry {
     FileCollectionFingerprinter getFingerprinter(FileNormalizationSpec spec);
