+++ subprojects/execution/src/main/java/org/gradle/internal/execution/fingerprint/FileCollectionSnapshotter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,17 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.fingerprint;
+package org.gradle.internal.execution.fingerprint;
 
-import org.gradle.api.tasks.FileNormalizer;
+import org.gradle.api.file.FileCollection;
+import org.gradle.internal.snapshot.FileSystemSnapshot;
 
 /**
- * Specifies criteria for selecting a {@link FileCollectionFingerprinter}
+ * Service for snapshotting {@link FileCollection}s.
  */
-public interface FileNormalizationSpec {
-    Class<? extends FileNormalizer> getNormalizer();
-
-    DirectorySensitivity getDirectorySensitivity();
+public interface FileCollectionSnapshotter {
+    /**
+     * Returns snapshots of the roots of a file collection.
+     */
+    FileSystemSnapshot snapshot(FileCollection fileCollection);
 }
